ac3strip: Replace all uses of copyint.
Now that we have a context struct, and the context struct has the
bitreader and bitwriter, we don't have to pass them around explicitly
any more. This makes the copy calls a bit cleaner, which is good because
there are a lot of them.

Also drop the "int" suffix because it seems slightly inaccurate.