ac3strip: hook up output, fix a couple bugs
Milestone: for the first time, we can parse the first audio block of the
first frame more or less correctly, i think, up to the mantissas.

Of course trying to parse the rest of the the frame blows right past the
first frame and into the second, so something is broken.

- Fix bitwriter

- Copy missing gainrng field