remember to set input/output handles for the flac process
Note: i'm not sure if i should be passing the parent's output and error
handles directly to the child like this. A comment on stackoverflow
suggests that that's a bad idea, and example code creates a pipe
for all three, but it seems to work. Besides, we don't expect any output
from the flac process.

https://stackoverflow.com/questions/19051769/createprocess-redirect-child-stdout-to-parent-stdout-in-c-windows#19844720