delay closing ifo0 until the end of main
We open title 0 to get info about the number of titles, and then close
it and open the title we really want. This ifoClose call in implicated
in some weird segfaults i've gotten on windows, so instead move it to
the end of the program, leaving ifo0 open until the program exits.